+++ packages/esm-patient-conditions-app/src/conditions/conditions-action-menu.scss
@@ -1,7 +1,5 @@
-.conditionsOverflowMenuLayer {
-  &:global(.cds--layer-two) {
-    height: 100%;
-  }
+.layer {
+  height: 100%;
 }
 
 .menuItem {
