+++ packages/esm-patient-conditions-app/translations/en.json
@@ -5,6 +5,8 @@
   "cancel": "Cancel",
   "checkFilters": "Check the filters above",
   "condition": "Condition",
+  "conditionDeleted": "Condition Deleted",
+  "conditionDeletedSuccessfully": "Condition has been deleted successfully",
   "conditionEditingError": "Error editing condition",
   "conditionNowVisible": "It is now visible on the Conditions page",
   "conditions": "Conditions",
@@ -14,12 +16,13 @@
   "currentStatus": "Current status",
   "dateOfOnset": "Date of onset",
   "delete": "Delete",
-  "deleteConditionTitle": "Are you sure you want to delete this condition?",
-  "deleteConfirmationText": "You can change conditions to \"Inactive\" using Edit mode.",
+  "deleteCondition": "Delete condition",
+  "deleteModalConfirmationText": "Are you sure you want to delete this condition?",
   "edit": "Edit",
+  "editCondition": "Edit a Condition",
   "endDate": "End date",
   "enterCondition": "Enter condition",
-  "errorDeleting": "Error deleting a condition",
+  "errorDeletingCondition": "Error deleting a condition",
   "inactive": "Inactive",
   "noConditionsToDisplay": "No conditions to display",
   "noResultsFor": "No results for",
